+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.9]
+
+### Changed
+
+- Standardized logging
+
+### Fixed
+
+- Added missing `continue` statement on main loop
+
 ## [0.0.8]
 
 ### Changed